test: relocate paste Enter regression test for test-size ratchet
