+++ crates/jcode-tui/src/tui/app/tests.rs
@@ -1,5 +1,4 @@
 #![cfg_attr(test, allow(clippy::clone_on_copy))]
-
 include!("tests/support_failover/part_01.rs");
 include!("tests/support_failover/part_02.rs");
 include!("tests/commands_accounts_01/part_01.rs");
@@ -44,6 +43,7 @@ include!("tests/smoothness_benchmark.rs");
 include!("tests/hotkey_feedback_e2e.rs");
 include!("tests/todo_card.rs");
 include!("tests/issue_496_input_routing.rs");
+include!("tests/issue_544_paste_enter.rs");
 include!("tests/issue_497_copy_ctrl_c.rs");
 
 #[test]
@@ -1740,35 +1740,4 @@ fn remote_clear_resets_provider_reported_context_usage() {
         .expect("remote /clear should succeed");
 
     assert_clear_usage_reset(&app);
-}
-
-#[test]
-fn bare_enter_immediately_after_paste_does_not_submit() {
-    // Windows Terminal / conhost sends a separate bare Enter key event after
-    // a bracketed paste ending with \n; it must not submit the chat (#544).
-    use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
-
-    let mut app = create_test_app();
-    crate::tui::app::input::handle_paste(&mut app, "hello world\n".to_string());
-    assert_eq!(
-        app.input,
-        "hello world\n".trim_end_matches('\n').to_owned() + "\n"
-    );
-
-    app.handle_key_press_event(KeyEvent::new(KeyCode::Enter, KeyModifiers::NONE))
-        .unwrap();
-    assert!(
-        !app.is_processing,
-        "Enter right after paste must not submit"
-    );
-    assert!(
-        !app.input.is_empty(),
-        "input should be preserved after paste"
-    );
-
-    // A later, human-timed Enter still submits.
-    crate::tui::app::input::paste_guard_expire_for_test();
-    app.handle_key_press_event(KeyEvent::new(KeyCode::Enter, KeyModifiers::NONE))
-        .unwrap();
-    assert!(app.input.is_empty(), "later Enter should submit normally");
 }
