+++ crates/jcode-tui/src/tui/app/tests/issue_544_paste_enter.rs
@@ -1,0 +1,31 @@
+// Issue #544: stray Enter after a bracketed paste must not submit.
+#[test]
+fn bare_enter_immediately_after_paste_does_not_submit() {
+    // Windows Terminal / conhost sends a separate bare Enter key event after
+    // a bracketed paste ending with \n; it must not submit the chat (#544).
+    use crossterm::event::{KeyCode, KeyEvent, KeyModifiers};
+
+    let mut app = create_test_app();
+    crate::tui::app::input::handle_paste(&mut app, "hello world\n".to_string());
+    assert_eq!(
+        app.input,
+        "hello world\n".trim_end_matches('\n').to_owned() + "\n"
+    );
+
+    app.handle_key_press_event(KeyEvent::new(KeyCode::Enter, KeyModifiers::NONE))
+        .unwrap();
+    assert!(
+        !app.is_processing,
+        "Enter right after paste must not submit"
+    );
+    assert!(
+        !app.input.is_empty(),
+        "input should be preserved after paste"
+    );
+
+    // A later, human-timed Enter still submits.
+    crate::tui::app::input::paste_guard_expire_for_test();
+    app.handle_key_press_event(KeyEvent::new(KeyCode::Enter, KeyModifiers::NONE))
+        .unwrap();
+    assert!(app.input.is_empty(), "later Enter should submit normally");
+}
